requirement.txt
ctransformers==0.2.5
flask
pypdf
sentence_transformers
python-dotenv
langchain==0.3.0
langchain-community==0.3.0
pinecone>=3.0.0
sentence-transformers>=2.2.2
pydantic>=2.7,<3
protobuf==4.25.3
-e .